Address review comment
